+++ internal/rsm/offload.go
@@ -27,6 +27,9 @@ const (
 	// FromCommitWorker indicates that the data store has been loaded by or
 	// offloaded from the commit worker.
 	FromCommitWorker
+	// FromApplyWorker indicates that the data store has been loaded by or
+	// offloaded from the apply worker.
+	FromApplyWorker
 	// FromSnapshotWorker indicates that the data store has been loaded by or
 	// offloaded from the snapshot worker.
 	FromSnapshotWorker
@@ -36,6 +39,7 @@ var fromNames = [...]string{
 	"FromNodeHost",
 	"FromStepWorker",
 	"FromCommitWorker",
+	"FromApplyWorker",
 	"FromSnapshotWorker",
 }
 
@@ -52,9 +56,11 @@ type OffloadedStatus struct {
 	offloadedFromNodeHost       bool
 	offloadedFromStepWorker     bool
 	offloadedFromCommitWorker   bool
+	offloadedFromApplyWorker    bool
 	offloadedFromSnapshotWorker bool
 	loadedByStepWorker          bool
 	loadedByCommitWorker        bool
+	loadedByApplyWorker         bool
 	loadedBySnapshotWorker      bool
 	DestroyedC                  chan struct{}
 }
@@ -87,6 +93,7 @@ func (o *OffloadedStatus) SetLoaded(from From) {
 	if o.offloadedFromNodeHost {
 		if from == FromStepWorker ||
 			from == FromCommitWorker ||
+			from == FromApplyWorker ||
 			from == FromSnapshotWorker {
 			plog.Panicf("loaded from %v after offloaded from nodehost", from)
 		}
@@ -97,6 +104,8 @@ func (o *OffloadedStatus) SetLoaded(from From) {
 		o.loadedByStepWorker = true
 	} else if from == FromCommitWorker {
 		o.loadedByCommitWorker = true
+	} else if from == FromApplyWorker {
+		o.loadedByApplyWorker = true
 	} else if from == FromSnapshotWorker {
 		o.loadedBySnapshotWorker = true
 	} else {
@@ -113,6 +122,8 @@ func (o *OffloadedStatus) SetOffloaded(from From) {
 		o.offloadedFromStepWorker = true
 	} else if from == FromCommitWorker {
 		o.offloadedFromCommitWorker = true
+	} else if from == FromApplyWorker {
+		o.offloadedFromApplyWorker = true
 	} else if from == FromSnapshotWorker {
 		o.offloadedFromSnapshotWorker = true
 	} else {
@@ -125,12 +136,16 @@ func (o *OffloadedStatus) SetOffloaded(from From) {
 		if !o.loadedByCommitWorker {
 			o.offloadedFromCommitWorker = true
 		}
+		if !o.loadedByApplyWorker {
+			o.offloadedFromApplyWorker = true
+		}
 		if !o.loadedBySnapshotWorker {
 			o.offloadedFromSnapshotWorker = true
 		}
 	}
 	if o.offloadedFromNodeHost &&
 		o.offloadedFromCommitWorker &&
+		o.offloadedFromApplyWorker &&
 		o.offloadedFromSnapshotWorker &&
 		o.offloadedFromStepWorker {
 		o.readyToDestroy = true