+++ internal/rsm/taskqueue.go
@@ -58,6 +58,16 @@ func (tq *TaskQueue) Add(task Task) {
 	tq.mu.Unlock()
 }
 
+// GetAll returns all tasks currently in the queue.
+func (tq *TaskQueue) GetAll() []Task {
+	tq.mu.Lock()
+	defer tq.mu.Unlock()
+	result := tq.tasks
+	tq.tasks = make([]Task, 0, initialTaskQueueCap)
+	tq.next = 0
+	return result
+}
+
 // Get returns a task from the queue if there is any.
 func (tq *TaskQueue) Get() (Task, bool) {
 	tq.mu.Lock()